@@ -1,86 +1,83 @@
 {
-    "type": "project",
-    "license": "proprietary",
-    "minimum-stability": "stable",
-    "prefer-stable": true,
-    "require": {
-        "php": ">=8.1",
-        "ext-ctype": "*",
-        "ext-iconv": "*",
-        "doctrine/dbal": "^3",
-        "doctrine/doctrine-bundle": "^2.15",
-        "doctrine/doctrine-migrations-bundle": "^3.4",
-        "doctrine/mongodb-odm-bundle": "^5.3",
-        "doctrine/orm": "^3.5",
-        "nelmio/cors-bundle": "^2.5",
-        "symfony/asset": "6.4.*",
-        "symfony/console": "6.4.*",
-        "symfony/dotenv": "6.4.*",
-        "symfony/flex": "^2",
-        "symfony/form": "6.4.*",
-        "symfony/framework-bundle": "6.4.*",
-        "symfony/mailer": "6.4.*",
-        "symfony/runtime": "6.4.*",
-        "symfony/security-bundle": "6.4.*",
-        "symfony/serializer": "6.4.*",
-        "symfony/twig-bundle": "6.4.*",
-        "symfony/validator": "6.4.*",
-        "symfony/yaml": "6.4.*",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
-    },
-    "config": {
-        "allow-plugins": {
-            "php-http/discovery": true,
-            "symfony/flex": true,
-            "symfony/runtime": true
-        },
-        "sort-packages": true
-    },
-    "autoload": {
-        "psr-4": {
-            "App\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "App\\Tests\\": "tests/"
-        }
-    },
-    "replace": {
-        "symfony/polyfill-ctype": "*",
-        "symfony/polyfill-iconv": "*",
-        "symfony/polyfill-php72": "*",
-        "symfony/polyfill-php73": "*",
-        "symfony/polyfill-php74": "*",
-        "symfony/polyfill-php80": "*",
-        "symfony/polyfill-php81": "*"
-    },
-    "scripts": {
-        "auto-scripts": {
-            "cache:clear": "symfony-cmd",
-            "assets:install %PUBLIC_DIR%": "symfony-cmd"
-        },
-        "post-install-cmd": [
-            "@auto-scripts"
-        ],
-        "post-update-cmd": [
-            "@auto-scripts"
-        ]
-    },
-    "conflict": {
-        "symfony/symfony": "*"
-    },
-    "extra": {
-        "symfony": {
-            "allow-contrib": false,
-            "require": "6.4.*"
-        }
-    },
-    "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^4.1",
-        "fakerphp/faker": "^1.24",
-        "symfony/maker-bundle": "^1.64",
-        "symfony/web-profiler-bundle": "6.4.*"
+  "name": "amelieoltz/ecoride",
+  "description": "EcoRide - application de covoiturage écoresponsable (Symfony 6.4, MySQL + MongoDB ODM)",
+  "type": "project",
+  "license": "proprietary",
+  "require": {
+    "php": ">=8.2",
+    "ext-ctype": "*",
+    "ext-iconv": "*",
+    "ext-mongodb": "*",
+    "doctrine/data-fixtures": "^1.7",
+    "doctrine/doctrine-bundle": "^2.18",
+    "doctrine/doctrine-fixtures-bundle": "^3.5",
+    "doctrine/doctrine-migrations-bundle": "^3.5",
+    "doctrine/mongodb-odm": "^2.3,<2.12",
+    "doctrine/mongodb-odm-bundle": "4.6.*",
+    "doctrine/orm": "^3.5",
+    "mongodb/mongodb": "1.19.*",
+    "nelmio/cors-bundle": "^2.6",
+    "phpdocumentor/reflection-docblock": "^5.6",
+    "phpstan/phpdoc-parser": "^2.3",
+    "platformsh/config-reader": "^2",
+    "symfony/asset": "6.4.*",
+    "symfony/console": "^6.4",
+    "symfony/flex": "^2.4",
+    "symfony/framework-bundle": "^6.4",
+    "symfony/mailer": "^6.4",
+    "symfony/property-access": "6.4.*",
+    "symfony/property-info": "6.4.*",
+    "symfony/runtime": "^6.4",
+    "symfony/security-bundle": "^6.4",
+    "symfony/serializer": "6.4.*",
+    "symfony/twig-bundle": "^6.4",
+    "symfony/validator": "^6.4",
+    "symfony/yaml": "^6.4",
+    "twig/extra-bundle": "^3.22"
+  },
+  "require-dev": {
+    "symfony/maker-bundle": "^1.60",
+    "symfony/web-profiler-bundle": "^6.4",
+    "fakerphp/faker": "^1.24",
+    "symfony/dotenv": "^6.4"
+  },
+  "autoload": {
+    "psr-4": {
+      "App\\": "src/"
     }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "App\\Tests\\": "tests/"
+    }
+  },
+  "scripts": {
+    "auto-scripts": {
+      "cache:clear": "symfony-cmd",
+      "assets:install %PUBLIC_DIR%": "symfony-cmd"
+    },
+    "post-install-cmd": [
+      "@auto-scripts"
+    ],
+    "post-update-cmd": [
+      "@auto-scripts"
+    ]
+  },
+  "extra": {
+    "symfony": {
+      "allow-contrib": true,
+      "require": "6.4.*"
+    }
+  },
+  "config": {
+    "optimize-autoloader": true,
+    "sort-packages": true,
+    "allow-plugins": {
+      "symfony/flex": true,
+      "symfony/runtime": true
+    },
+    "platform": {
+      "ext-mongodb": "1.19.4"
+    }
+  }
 }
