@@ -1,10 +1,11 @@
 <?php
-ini_set('display_errors', '1');
-error_reporting(E_ALL);
+
 use App\Kernel;
 
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
 return function (array $context) {
+    require_once dirname(__DIR__).'/config/bootstrap_platform.php';
+
     return new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
 };
